LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/declared/getAll-shorthand-expected.txt


PASS StylePropertyMap.getAll() with a shorthand property set explicitly in css rule returns a base CSSStyleValue
PASS StylePropertyMap.getAll() with a shorthand property that is partially in css rule returns empty list
